web3torrent: rename magnet.ts -> url.ts
